% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinFlip.R
\name{CoinFlip}
\alias{CoinFlip}
\title{CoinFlip Function}
\usage{
CoinFlip()
}
\value{
returns a character string
}
\description{
This function flips a coin
}
\examples{
CoinFlip()

}
\keyword{coin}
\keyword{flip}
