% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvenOrOdd.R
\name{EvenOrOdd}
\alias{EvenOrOdd}
\title{EvenOrOdd Function}
\usage{
EvenOrOdd(number)
}
\arguments{
\item{numeric}{The number you put in.}
}
\value{
returns a character string
}
\description{
This function tells you if your number is even or odd
}
\examples{
EvenOrOdd(numbers)
EvenOrOdd(6)
}
\keyword{number}
